Do not set contentScale based on UIWindow.
For current devices with 'retina' resolution (2 pixels/1 point),
the contentsScale of a layer needs to be set to 2.0. The mainScreen
singleton has the correct scale factor, but the UIWindow.contentsScaleFactor
is not always correct. Setting the layer's contentsScale based on
the UIWindow can result in low-res rendering on retina displays.

The window's contentScaleFactor does not always contain the correct